pyproject.toml
[project]
name = "lesson-4"
version = "0.1.0"
description = ""
authors = [
    {name = "Oleksandr",email = "alexandrenko3006@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

